move fs watcher to a separate module
